requirements.txt
ArducamSDK
arducam_config_parser
arducam_rgbir_debayer
